chore: add some comments / move some code
picked from #10910
